+++ tests/openassetio-core-c/private/hostApi/ManagerTest.cpp
@@ -39,7 +39,7 @@ struct MockManagerInterface : trompeloeil::mock_interface<managerApi::ManagerInt
   IMPLEMENT_CONST_MOCK0(identifier);
   IMPLEMENT_CONST_MOCK0(displayName);
   IMPLEMENT_CONST_MOCK0(info);
-  IMPLEMENT_MOCK1(initialize);
+  IMPLEMENT_MOCK2(initialize);
   IMPLEMENT_CONST_MOCK3(managementPolicy);
 };
 /**